로그인 후 이동 부분 구현

+++ src/App.js
@@ -14,17 +14,29 @@ function App() { // 페이지로 보여주는 부분 반드시 필요함. 임의
       setStructure(localStorage.getItem('structure'));
     }
   }, []);
+
+  function navTo(component, isRedirect) { // 원하는 위치로 리다이렉트 or 컴포넌트를 보여주기 or 웰컴페이지로 돌아가기
+    if (localStorage.getItem('logininfo') && isRedirect) { // 리다이렉트를 하고 싶을 때
+      return <Navigate to={component} replace />
+    }
+    else if (localStorage.getItem('logininfo') && !isRedirect) { // 컴포넌트를 반환하고 싶을 때
+      return component;
+    }
+    return <Navigate to={'/welcome'} replace />
+  }
+
   return (
     <div className="App">
       <BrowserRouter> {/* 페이지 이동하는 부분 가급적 건드리지 마시오. */}
         <Routes>
-          <Route path="*" element={<Navigate to={'/map'} replace />} />
-          <Route path="/map" element={<>
+          <Route path="*" element={navTo('/map', true)} />
+          <Route path="/map" element={<> {/* 이쪽은 로그인 변환용으로 일단 냅둠 */}
             <Nav setStructure={setStructure} />
             <Map structure={structure} floor={floor} setFloor={setFloor} setStructure={setStructure} />
           </>} />
-          <Route path="/teach" element={<><Nav setStructure={setStructure} /><Teach /></>} />
+          <Route path="/teach" element={navTo(<><Nav setStructure={setStructure} /><Teach /></>, false)} />
           <Route path="/login" element={<><Login /></>} />
+          <Route path="/welcome" element={<><Welcome /></>} />
         </Routes>
       </BrowserRouter>
     </div>
@@ -68,16 +80,19 @@ function Nav({ setStructure }) { // 네비게이션 바
       {!logined ? <>
         <a href='/login'><button>로그인</button></a>&nbsp; { /* 로그인 페이지 이동하는 버튼 */}
         <button onClick={e => { // 회원가입하는 버튼 (임시) 추후 페이지 이동
-          setLogined(e => '홍길동');
-          localStorage.setItem('logininfo', '홍길동');
+          setLogined(e => 'Guest');
+          localStorage.setItem('logininfo', 'Guest');
           localStorage.setItem('logintime', time.toLocaleTimeString());
         }}>회원가입</button></>
         : <>
           <div className="clientInfo">
             <span>{logined}</span>&nbsp;&nbsp;
             <button onClick={e => { // 로그아웃하는 버튼
-              setLogined(e => false);
-              localStorage.clear(); // 로그아웃 시 저장된 모든 정보 삭제
+              if (window.confirm("정말로?")) { // 로그아웃할 건지 재확인
+                setLogined(e => false);
+                localStorage.clear(); // 로그아웃 시 저장된 모든 정보 삭제
+                window.location.reload(); // 페이지 다시 로드
+              }
             }}>로그아웃</button>
           </div>
         </>
@@ -105,7 +120,19 @@ function Teach() { //선생님 페이지
 }
 
 function Login() { //로그인 화면 페이지
+  const time = new Date();
   return <div className="login">
-    login
+    <button onClick={e => {
+      localStorage.setItem('logininfo', 'Guest');
+      localStorage.setItem('logintime', time.toLocaleTimeString());
+      window.location.href = '/map'
+    }}>게스트 로그인</button>
+  </div>;
+}
+
+function Welcome() {
+  return <div className="welcome">
+    welcome
+    <button onClick={e => window.location.href = '/login'}>로그인 하러가기</button>
   </div>;
 }
